Create README.md
Based on the content from the Tiloid website, here's a suggested structure for a README.md file:

# Today I Learned for Programmers

## About
This is a collection of short programming-related articles and tutorials, curated by the Tiloid community. Each post covers a specific concept, tool, or technique that the author learned recently and found useful.

The goal of this project is to provide a readily accessible knowledge base for programmers of all skill levels to learn from the experiences of others.

## Featured Posts
Here are some of the most popular or recent posts on Tiloid:

1. **[The Ultimate Guide to README.md](https://tiloid.com/readme-the-ultimate-guide)** by Naveen Sahu
2. **[Converting a String to Enum in TypeScript](https://tiloid.com/converting-a-string-to-enum)** by Alba Silvente Fuentes
3. **[Recursively Grep All Directories and Subdirectories in Linux](https://tiloid.com/recursively-grep-all-directories-and-subdirectories)** by Annie
4. **[Setting the Default Node Version using NVM](https://tiloid.com/setting-the-default-node-version-using-nvm)** by Michael Fienen
5. **[Adding Multiple Classes to a React Component](https://tiloid.com/adding-multiple-classes-to-a-react-component)** by Caio Teixeira

## Contributing
If you have a programming-related tip, trick, or lesson you'd like to share, feel free to submit a new post to Tiloid. The process is simple:

1. Create a new Markdown file in the appropriate category folder.
2. Follow the established formatting and structure.
3. Submit a pull request to the Tiloid repository.

Your contribution will be reviewed and, if accepted, added to the collection.

## License
This project is licensed under the [MIT License](LICENSE). Feel free to use the content for personal or commercial purposes, but please provide attribution to the original authors.